# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API",
        "description": "Documentation de l’API MusicBox (Artistes, Albums, Chansons)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API local"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Lister les albums",
                "description": "Retourne la liste des albums avec pagination et filtres optionnels",
                "operationId": "735609843dcfb89ee4e12e340eddca25",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filtrer par titre de l'album",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "artist",
                        "in": "query",
                        "description": "Filtrer par nom d'artiste",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filtrer par année",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée d'albums",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Album"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Créer un album",
                "operationId": "2252aa8fd63b7d2fa1203c1735943d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Album"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album créé avec succès"
                    }
                }
            }
        },
        "/api/albums/{album}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Voir un album",
                "operationId": "1438c74c20588b81ffcdbc54e9b5258e",
                "parameters": [
                    {
                        "name": "album",
                        "in": "path",
                        "description": "ID de l'album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail d'un album"
                    }
                }
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Mettre à jour un album",
                "operationId": "0c94feaa993026ec58a585ef8ded6778",
                "parameters": [
                    {
                        "name": "album",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Album"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Supprimer un album",
                "operationId": "1ea877ac4c6df523f3c957cf161bc142",
                "parameters": [
                    {
                        "name": "album",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Album supprimé"
                    }
                }
            }
        },
        "/api/albums/{album}/songs": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Lister les chansons d'un album",
                "operationId": "77dda7491f8f4a514fe2deb5794aab51",
                "parameters": [
                    {
                        "name": "album",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chansons de l'album"
                    }
                }
            }
        },
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get list of artists with pagination and filters",
                "operationId": "2c8aed1387cca1c65d98da7ee11ac926",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by artist name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Filter by genre (comma-separated for multiple)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Filter by country",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of artists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Artist"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Create a new artist",
                "operationId": "83d2888bdc6fac66b9a7cc81c4c950fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Artist"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artist created successfully"
                    }
                }
            }
        },
        "/api/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get artist details by ID",
                "operationId": "d0b68c06b8bc310fbe24c6d3ea80dac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Update an existing artist",
                "operationId": "5a5632ea389f9a05336f8d035f5a44c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Artist"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artist updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Delete an artist",
                "operationId": "4f6d8bd2c8b4b738af7589b72348ac20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artist deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Album": {
                "title": "Album",
                "required": [
                    "title",
                    "year",
                    "artist_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "title": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "artist_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Artist": {
                "title": "Artist",
                "required": [
                    "name",
                    "genre",
                    "country"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "genre": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}